refactor: UserAccountInformation
