+++ model/User.py
@@ -1,25 +1,24 @@
-from model import UserInformation
+from model.UserPersonalInformation import UserPersonalInformation
 from model.SurveyAnswers import SurveyAnswers
+from model.UserAccountInformation import UserAccountInformation
 
 
 class User:
     def __init__(
         self,
-        username: str,
-        password: str,
-        information: UserInformation = None,
+        account_information: UserAccountInformation,
+        personal_information: UserPersonalInformation = None,
         answers: SurveyAnswers = None,
     ):
-        self.username = username
-        self.password = password
-        self.information = information
+        self.account_information = account_information
+        self.personal_information = personal_information
         self.answers = answers
 
     def __eq__(self, other):
         if not isinstance(other, User):
             return False
 
-        return self.username == other.username
+        return self.account_information.username == other.account_information.username
 
     def __str__(self):
         return f"{self.username}"
